requirements.txt
Flask
opencv-python
numpy
Pillow
tensorflow
transformers
huggingface_hub
requests